refactor(Frames): 볼링게임 이름을 frames로 변경
