+++ src/main/java/bowling/controller/BowlingGameController.java
@@ -1,8 +1,8 @@
 package bowling.controller;
 
-import bowling.domain.BowlingGame;
 import bowling.domain.Pin;
 import bowling.domain.PlayerName;
+import bowling.domain.frame.Frames;
 import bowling.domain.scorestrategy.RandomScoreStrategy;
 import bowling.view.InputView;
 import bowling.view.OutputView;
@@ -11,13 +11,15 @@ public class BowlingGameController {
 
     public static void main(String[] args) {
         PlayerName playerName = new PlayerName(InputView.askPlayerName());
-        BowlingGame bowlingGame = new BowlingGame();
-        OutputView.printStart(bowlingGame, playerName);
+        Frames frames = new Frames();
+        OutputView.printStart(frames, playerName);
 
-        while (!bowlingGame.isFinished()) {
-            Pin now = bowlingGame.bowl(new RandomScoreStrategy());
-            OutputView.print(bowlingGame, playerName, now);
+        while (!frames.isFinished()) {
+            Pin now = frames.bowl(new RandomScoreStrategy());
+            OutputView.print(frames, playerName, now);
         }
 
     }
+
+
 }
